refactor(macos): finish AppModel god-object extraction (Session/Library/Favorites/AlbumDetail/Window)
Extract the remaining cohesive method domains out of AppModel.swift into
focused `extension AppModel` files, completing the god-object split.
AppModel.swift drops 2,667 -> 1,392 lines (original was 7,106; -80% overall),
leaving essentially the @Observable stored-state surface plus init.

New extension files:
- AppModel+Session.swift     login / restore / logout, network retry, auth-error handling
- AppModel+Library.swift     library / playlist / genre loading, pagination, items-query fetch helpers
- AppModel+Favorites.swift   favorite + played toggles and reads
- AppModel+AlbumDetail.swift album-detail loading + album-row actions
- AppModel+Window.swift      Mini Player control, queue-end autoplay toggle, full-window restore

Two strays folded into existing extensions: enqueueDownload -> +Downloads,
addToPlaylist -> +PlaylistTracks.

Stored properties stay in AppModel.swift per the @Observable contract; only the
private members actually referenced cross-file by a moved method were promoted to
internal (page-size constants, ambient-palette caches, hasAttemptedRestore, the
mini-player / autoplay UserDefaults keys). Pure relocation, no behavior change.

Clean rebuild green; 889 tests pass.
