+++ macos/Sources/Lyrebird/AppModel.swift
@@ -458,19 +458,19 @@ final class AppModel {
     /// First-paint size for library lists. Tuned smaller than subsequent
     /// pages so the grid renders quickly on login; raising this increases
     /// time-to-first-paint without measurable benefit.
-    private let libraryInitialPageSize: UInt32 = 100
+    let libraryInitialPageSize: UInt32 = 100
     /// Follow-up page size for library lists. Larger than the initial page
     /// to keep round-trip count down once the user has committed to scrolling.
-    private let libraryPageSize: UInt32 = 200
+    let libraryPageSize: UInt32 = 200
     /// Initial page size for recently-played on the Home screen.
-    private let recentlyPlayedInitialPageSize: UInt32 = 20
+    let recentlyPlayedInitialPageSize: UInt32 = 20
     /// Page size when walking `playlist_tracks` to completion.
-    private let playlistPageSize: UInt32 = 200
+    let playlistPageSize: UInt32 = 200
     /// Hard cap on total tracks pulled from a single playlist to keep a
     /// pathological 50k-track playlist from holding the UI hostage. Callers
     /// that hit this see up to this many tracks; beyond that the rest is
     /// silently dropped. Easy to raise once a real use case complains.
-    private let playlistSafetyCap: Int = 5000
+    let playlistSafetyCap: Int = 5000
     /// Page size for the "Show all results" affordance in search.
     let searchPageSize: UInt32 = 50
 
@@ -682,7 +682,7 @@ final class AppModel {
     /// directly — we mirror the value into `miniPlayerAlwaysOnTop` and write
     /// through on each change, the same JSON/UserDefaults bridge idiom the
     /// pinned-stations store uses.
-    private static let miniPlayerAlwaysOnTopKey = "miniPlayer.alwaysOnTop"
+    static let miniPlayerAlwaysOnTopKey = "miniPlayer.alwaysOnTop"
 
     /// Whether the detached Mini Player window is currently open.
     /// `LyrebirdApp` observes this and drives `openWindow` / `dismissWindow`
@@ -702,27 +702,12 @@ final class AppModel {
     /// stored value and the live window level stay consistent.
     var miniPlayerAlwaysOnTop: Bool = UserDefaults.standard.bool(forKey: AppModel.miniPlayerAlwaysOnTopKey)
 
-    /// Toggle the Mini Player window. Wired to the ⌘⌥P menu command; flipping
-    /// the flag is enough — `LyrebirdApp` translates the change into the
-    /// matching `openWindow` / `dismissWindow` call for the `mini-player`
-    /// scene.
-    func toggleMiniPlayer() {
-        isMiniPlayerVisible.toggle()
-    }
-
-    /// Set + persist the always-on-top preference. `MiniPlayerWindowConfigurator`
-    /// reads `miniPlayerAlwaysOnTop` and re-applies the window level on change.
-    func setMiniPlayerAlwaysOnTop(_ on: Bool) {
-        miniPlayerAlwaysOnTop = on
-        UserDefaults.standard.set(on, forKey: AppModel.miniPlayerAlwaysOnTopKey)
-    }
-
     // MARK: - Autoplay (queue end)
 
     /// UserDefaults key for the persisted "autoplay similar music when the
     /// queue ends" preference. Same `@Observable` → UserDefaults bridge as
     /// the mini-player flag above.
-    private static let autoplayWhenQueueEndsKey = "queue.autoplayWhenQueueEnds"
+    static let autoplayWhenQueueEndsKey = "queue.autoplayWhenQueueEnds"
 
     /// Whether playback should extend with an Instant Mix of similar music
     /// when the user-added queue and its source tail run dry. Default **on**
@@ -732,59 +717,6 @@ final class AppModel {
     /// launches; read through `autoplayWhenQueueEndsDefault` so an unset key
     /// resolves to `true` rather than `bool(forKey:)`'s `false`.
     var autoplayWhenQueueEnds: Bool = AppModel.autoplayWhenQueueEndsDefault()
-
-    /// Resolve the persisted autoplay flag, defaulting to `true` when the key
-    /// has never been written. `UserDefaults.bool(forKey:)` returns `false`
-    /// for a missing key, which would silently invert this feature's "default
-    /// on" contract, so we probe for the object first.
-    private static func autoplayWhenQueueEndsDefault() -> Bool {
-        guard UserDefaults.standard.object(forKey: autoplayWhenQueueEndsKey) != nil else {
-            return true
-        }
-        return UserDefaults.standard.bool(forKey: autoplayWhenQueueEndsKey)
-    }
-
-    /// Set + persist the autoplay-at-queue-end preference. Wired to the queue
-    /// header toggle; `handleTrackEnded` reads `autoplayWhenQueueEnds` when
-    /// the queue runs dry to decide whether to extend with an Instant Mix or
-    /// stop.
-    func setAutoplayWhenQueueEnds(_ on: Bool) {
-        autoplayWhenQueueEnds = on
-        UserDefaults.standard.set(on, forKey: AppModel.autoplayWhenQueueEndsKey)
-    }
-
-    /// Close the Mini Player and bring the full window forward, honouring the
-    /// "closing returns to full window" contract. Used by the mini player's
-    /// settings-menu and hover "return" affordances. Clearing the flag lets
-    /// `LyrebirdApp` dismiss the scene; activating the app raises the main
-    /// `WindowGroup` window back to the foreground.
-    func returnToFullWindow() {
-        isMiniPlayerVisible = false
-        NSApplication.shared.activate(ignoringOtherApps: true)
-    }
-
-    /// Drill the **main** window to a detail `Route` from the detached Mini
-    /// Player, raising the full window first so the navigation is visible.
-    ///
-    /// The Mini Player floats in its own borderless window (optionally
-    /// always-on-top), so a bare `navigate(to:)` would push onto the main
-    /// window's `NavigationStack` while that window stays buried behind
-    /// everything else — the user would tap album art and see nothing move.
-    /// Activating the app raises the main `WindowGroup` window back to the
-    /// foreground the same way `returnToFullWindow` does, *then* we drill.
-    /// Unlike `returnToFullWindow` this intentionally leaves
-    /// `isMiniPlayerVisible` untouched: clicking through to a detail page is
-    /// not a request to dismiss the mini player, so an always-on-top widget
-    /// keeps floating over the now-foregrounded detail view (the Apple Music /
-    /// Spotify mini-widget contract).
-    ///
-    /// Routing through one seam (rather than letting `MiniPlayerView` poke
-    /// `navPath` + `NSApp` itself) keeps the activate-then-navigate ordering in
-    /// a single testable place and matches `openLyrics` / `navigate(to:)`.
-    func openInMainWindowFromMiniPlayer(_ route: Route) {
-        NSApplication.shared.activate(ignoringOtherApps: true)
-        navigate(to: route)
-    }
 
     // MARK: - Command Palette (⌘K)
 
@@ -900,915 +832,12 @@ final class AppModel {
         self.audio.offlinePlaybackEnabled = supportsDownloads
     }
 
-    // MARK: - Network
-
-    /// Re-evaluates network reachability and, if a session exists, kicks off a
-    /// library refetch. Wired to the offline banner's `Retry` button.
-    func retryNetwork() {
-        network.retry()
-        guard session != nil else { return }
-        Task { await refreshLibrary() }
-    }
-
-    /// Clears the server-reachability failure counter and retries the library
-    /// fetch. Wired to the server-unreachable banner's `Retry` button.
-    /// Resetting up-front means the banner disappears while the user waits;
-    /// if the refetch fails again, the error flow in `refreshLibrary` will
-    /// re-accumulate failures and the banner will come back.
-    func retryServer() {
-        serverReachability.reset()
-        guard session != nil else { return }
-        Task { await refreshLibrary() }
-    }
-
-    // MARK: - Session
-
-    func login(url: String, username: String, password: String) async {
-        isLoggingIn = true
-        defer { isLoggingIn = false }
-        do {
-            let session = try await Task.detached(priority: .userInitiated) { [core] in
-                try core.login(url: url, username: username, password: password)
-            }.value
-            self.session = session
-            self.serverURL = url
-            self.username = username
-            self.errorMessage = nil
-            startPolling()
-            await refreshLibrary()
-            await refreshDownloads()
-        } catch {
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .login)
-        }
-    }
-
-    /// Rehydrate the previous session from on-disk settings + the keychain
-    /// token. Called once from `RootView.task` on cold start. No-ops when the
-    /// core has nothing to restore (first launch, post-logout, etc.); in that
-    /// case `RootView` falls through to `LoginView`.
-    ///
-    /// Silent on errors: the core's `resume_session` is best-effort, so if the
-    /// local state is inconsistent we log and let the user sign in again
-    /// rather than blocking the app. Library fetches against the restored
-    /// session go through the regular `handleAuthError` flow, so a 401 on the
-    /// first call surfaces the auth-expired sheet just like a mid-session
-    /// expiry. Silent reauth is the rest of #440.
-    func attemptRestoreSession() async {
-        // Run at most once per AppModel lifetime. `hasAttemptedRestore`
-        // flips the first time this runs so re-renders of `RootView` that
-        // re-fire `.task` don't repeat the restore pass.
-        guard !hasAttemptedRestore else { return }
-        hasAttemptedRestore = true
-        guard session == nil else {
-            isRestoringSession = false
-            return
-        }
-        defer { isRestoringSession = false }
-        do {
-            let restored = try await Task.detached(priority: .userInitiated) { [core] in
-                try core.resumeSession()
-            }.value
-            guard let session = restored else { return }
-            self.session = session
-            self.serverURL = session.server.url
-            self.username = session.user.name
-            self.errorMessage = nil
-            startPolling()
-            await refreshLibrary()
-            await refreshDownloads()
-        } catch {
-            // Best-effort: leave `session == nil` so RootView renders LoginView.
-            // No banner — the user sees the login form, which is already the
-            // recovery path, and the library refetch after a manual sign-in
-            // will noisily surface any persistent server problem.
-            Log.auth.error("attemptRestoreSession failed: \(error.localizedDescription, privacy: .public)")
-        }
-    }
-
     /// Internal guard for `attemptRestoreSession` — the restore pass should
     /// run exactly once per app lifetime. Separate from `isRestoringSession`
     /// so the UI flag can be flipped without gating re-entry, and vice-versa.
-    private var hasAttemptedRestore = false
-
-    func logout() {
-        audio.stop()
-        // core.logout() does a blocking POST /Sessions/Logout; fire it off
-        // the main actor so the UI clears immediately and the user isn't
-        // staring at a stalled window for the network round-trip.
-        Task.detached(priority: .userInitiated) { [core] in
-            try? core.logout()
-        }
-        session = nil
-        imageURLCache.removeAll()
-        albums = []
-        artists = []
-        tracks = []
-        playlists = []
-        playlistLibraryId = nil
-        albumTracks = [:]
-        playlistTracks = [:]
-        currentPlaylistTracks = []
-        pendingPlaylistRemoval = nil
-        playlistDescriptions = [:]
-        sidebarEditingPlaylistId = nil
-        sidebarEditingDraft = ""
-        sidebarCopyingPlaylistIds = []
-        playlistPendingDelete = nil
-        artistTopTracks = [:]
-        artistSimilarCache = [:]
-        artistPlaylistsCache = [:]
-        artistAlbumsCache = [:]
-        artistDetailCache = [:]
-        resolvedNameCache = [:]
-        ambientPaletteCache.removeAll()
-        ambientPaletteTasks.values.forEach { $0.cancel() }
-        ambientPaletteTasks.removeAll()
-        recentlyPlayed = []
-        forYou = []
-        genresToExplore = []
-        browseGenres = []
-        jumpBackIn = []
-        recentlyAdded = []
-        recentlyAddedDates = [:]
-        quickPicks = []
-        quickPicksPlayCounts = [:]
-        favoriteAlbumsAll = []
-        favoriteAlbumsVisible = []
-        favoriteArtists = []
-        recentlyDiscoveredArtists = []
-        rediscover = []
-        suggestions = []
-        searchResults = nil
-        searchQuery = ""
-
-        searchPageResults = [:]
-        searchPageQuery = ""
-        searchPageActiveQuery = ""
-        activeSearchScope = .all
-        searchPageTotal = 0
-        searchPageLoaded = 0
-        searchPageExhausted = false
-        isLoadingFullSearch = false
-
-        currentTrackPeople = []
-        currentTrackPeopleForId = nil
-        currentLyrics = nil
-        currentLyricsForId = nil
-        sessionPlayHistory = []
-        // Clear the in-memory download snapshot. On-disk files are intentionally
-        // NOT removed — they're keyed by track id and rehydrate via
-        // `refreshDownloads()` on the next sign-in. See #819.
-        downloadStateById = [:]
-        downloads = []
-        downloadStats = nil
-        downloadsInFlight = []
-        resetPaginationState()
-        stopPolling()
-    }
-
-    /// Drop the stored access token (keychain + in-memory session) without
-    /// clearing the remembered server URL / username, so the user can re-auth
-    /// against the same server by re-entering only their password. Called
-    /// when the user taps "Sign in" on the auth-expired sheet. Note: the
-    /// caller still owns toggling `authExpired` off and nilling `session`.
-    ///
-    /// Unlike `logout`, this goes through the core's `forget_token` which
-    /// keeps `last_server_url` / `last_username` on disk for the login-form
-    /// prefill and only drops the credential store token plus the id settings
-    /// that key into it. So a subsequent `attemptRestoreSession` on next launch
-    /// short-circuits to `None` (safe), and the form is pre-populated.
-    func forgetToken() {
-        audio.stop()
-        try? core.forgetToken()
-        albums = []
-        artists = []
-        tracks = []
-        playlists = []
-        playlistLibraryId = nil
-        albumTracks = [:]
-        playlistTracks = [:]
-        currentPlaylistTracks = []
-        pendingPlaylistRemoval = nil
-        playlistDescriptions = [:]
-        sidebarEditingPlaylistId = nil
-        sidebarEditingDraft = ""
-        sidebarCopyingPlaylistIds = []
-        playlistPendingDelete = nil
-        artistTopTracks = [:]
-        artistSimilarCache = [:]
-        artistPlaylistsCache = [:]
-        artistAlbumsCache = [:]
-        artistDetailCache = [:]
-        resolvedNameCache = [:]
-        ambientPaletteCache.removeAll()
-        ambientPaletteTasks.values.forEach { $0.cancel() }
-        ambientPaletteTasks.removeAll()
-        recentlyPlayed = []
-        forYou = []
-        genresToExplore = []
-        browseGenres = []
-        jumpBackIn = []
-        recentlyAdded = []
-        recentlyAddedDates = [:]
-        quickPicks = []
-        quickPicksPlayCounts = [:]
-        favoriteAlbumsAll = []
-        favoriteAlbumsVisible = []
-        favoriteArtists = []
-        recentlyDiscoveredArtists = []
-        rediscover = []
-        suggestions = []
-        searchResults = nil
-        searchQuery = ""
-
-        searchPageResults = [:]
-        searchPageQuery = ""
-        searchPageActiveQuery = ""
-        activeSearchScope = .all
-        searchPageTotal = 0
-        searchPageLoaded = 0
-        searchPageExhausted = false
-        isLoadingFullSearch = false
-
-        currentTrackPeople = []
-        currentTrackPeopleForId = nil
-        currentLyrics = nil
-        currentLyricsForId = nil
-        sessionPlayHistory = []
-        resetPaginationState()
-        stopPolling()
-    }
-
-    /// Clear all pagination counters and in-flight flags. Kept in one place
-    /// so the two clear-the-session entry points (`logout`, `forgetToken`)
-    /// stay in sync.
-    private func resetPaginationState() {
-        albumsTotal = 0
-        artistsTotal = 0
-        tracksTotal = 0
-        playlistsTotal = 0
-        recentlyPlayedTotal = 0
-        searchResultsTotal = 0
-        isLoadingMoreAlbums = false
-        isLoadingMoreArtists = false
-        isLoadingMoreTracks = false
-        isLoadingMorePlaylists = false
-        isLoadingMoreSearch = false
-    }
-
-    /// Flag the session as expired. The UI surfaces this via the auth-expired
-    /// modal in `MainShell`. Idempotent — second hits within a session are
-    /// no-ops while the prompt is still visible.
-    func markAuthExpired() {
-        guard !authExpired else { return }
-        audio.stop()
-        stopPolling()
-        authExpired = true
-    }
-
-    /// Inspect an error from a core call and, if it's the core's
-    /// `NotAuthenticated` / `Auth` variant (both meaning the token's dead
-    /// or never existed), mark the session expired and return `true` so the
-    /// caller knows to skip its generic error surfacing.
-    ///
-    /// Post-BATCH-24 the Rust `LyrebirdError` is a typed enum split by HTTP
-    /// class — 401 responses surface as `Auth`, the retry-layer fallback is
-    /// `AuthExpired`, and a missing token is `NotAuthenticated` — so we can
-    /// match variants directly instead of parsing the Display message.
-    ///
-    /// Call-sites that do NOT match auth go on to call
-    /// `LyrebirdErrorPresenter.message(for:context:)` (see #351) to turn the
-    /// raw Display string into localized banner copy.
-    ///
-    /// `internal` (not `private`) so AppModel extension files in other source
-    /// files (e.g. `AppModel+Downloads.swift`) can route their FFI failures
-    /// through the same auth-expiry interception.
-    func handleAuthError(_ error: Error) -> Bool {
-        guard let err = error as? LyrebirdError else { return false }
-        switch err {
-        case .NotAuthenticated, .Auth, .AuthExpired:
-            markAuthExpired()
-            return true
-        default:
-            return false
-        }
-    }
-
-    // MARK: - Library
-
-    func refreshLibrary() async {
-        isLoadingLibrary = true
-        defer { isLoadingLibrary = false }
-        // Fetch albums, artists, tracks, and playlists in parallel. Previously
-        // the album/artist calls were sequential, doubling time-to-first-paint
-        // on every fresh session; `async let` lets all round-trips overlap.
-        // Playlists are wired in alongside so switching to the Playlists chip
-        // doesn't trigger a first-paint spinner. The smaller
-        // `libraryInitialPageSize` (100 vs. the old 200) is a further
-        // first-paint win — the grid fills the viewport with 100 and the
-        // per-tab `loadMore*` paths take over when the user scrolls.
-        //
-        // Playlists go through their own try/catch because the library id
-        // resolution can fail independently (no playlist library on the
-        // server, or an error from a hypothetical future `core.libraries()`)
-        // and we don't want that to sink the albums/artists/tracks fetch.
-        async let albumsPage = Task.detached(priority: .userInitiated) { [core, libraryInitialPageSize] in
-            try core.listAlbums(offset: 0, limit: libraryInitialPageSize)
-        }.value
-        async let artistsPage = Task.detached(priority: .userInitiated) { [core, libraryInitialPageSize] in
-            try core.listArtists(offset: 0, limit: libraryInitialPageSize)
-        }.value
-        async let tracksPage = Task.detached(priority: .userInitiated) { [core, libraryInitialPageSize] in
-            try core.listTracks(musicLibraryId: nil, offset: 0, limit: libraryInitialPageSize)
-        }.value
-        async let playlistsResult: Void = refreshPlaylists()
-        // Each loader gets its own try/catch so one failure (typically a
-        // transient 5xx on a single endpoint) doesn't drop the other two.
-        // Before, the tuple-destructure `try await (a, b, c)` cancelled the
-        // assignments for all three on any single error — Library rendered
-        // empty even when two of the three endpoints succeeded.
-        var anySucceeded = false
-        do {
-            let albums = try await albumsPage
-            self.albums = albums.items
-            self.albumsTotal = albums.totalCount
-            anySucceeded = true
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .libraryLoad)
-        }
-        do {
-            let artists = try await artistsPage
-            self.artists = artists.items
-            self.artistsTotal = artists.totalCount
-            anySucceeded = true
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .libraryLoad)
-        }
-        do {
-            let tracks = try await tracksPage
-            self.tracks = tracks.items
-            self.tracksTotal = tracks.totalCount
-            anySucceeded = true
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .libraryLoad)
-        }
-        if anySucceeded {
-            serverReachability.noteSuccess()
-        }
-        _ = await playlistsResult
-        await refreshRecentlyPlayed()
-        await refreshForYou()
-        await refreshGenresToExplore()
-        await refreshBrowseGenres()
-        // Home screen carousels (#49 / #51–#55). Kicked off after the main
-        // library so first paint isn't blocked on these secondary shelves.
-        // Each of these is best-effort — empty or errored rows just hide in
-        // the Home layout.
-        await refreshJumpBackIn()
-        await refreshRecentlyAdded()
-        await refreshQuickPicks()
-        await refreshFavoriteAlbums()
-        await refreshFavoriteArtists()
-        await refreshRecentlyDiscoveredArtists()
-        await refreshRediscover()
-        await refreshSuggestions()
-    }
-
-    /// Fetch the next page of albums and append to `albums`. No-op when a
-    /// page is already in flight or when the local count has caught up to
-    /// `albumsTotal`. Called from `LibraryView`'s near-end `.onAppear`
-    /// trigger — see `LibraryView.swift`.
-    func loadMoreAlbums() async {
-        guard !isLoadingMoreAlbums else { return }
-        guard albumsTotal == 0 || albums.count < Int(albumsTotal) else { return }
-        isLoadingMoreAlbums = true
-        defer { isLoadingMoreAlbums = false }
-        let offset = UInt32(albums.count)
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core, libraryPageSize] in
-                try core.listAlbums(offset: offset, limit: libraryPageSize)
-            }.value
-            self.albums.append(contentsOf: page.items)
-            self.albumsTotal = page.totalCount
-            serverReachability.noteSuccess()
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .libraryLoad)
-        }
-    }
-
-    /// Fetch the next page of artists and append to `artists`. Mirror of
-    /// `loadMoreAlbums` — see its docs for the trigger contract.
-    func loadMoreArtists() async {
-        guard !isLoadingMoreArtists else { return }
-        guard artistsTotal == 0 || artists.count < Int(artistsTotal) else { return }
-        isLoadingMoreArtists = true
-        defer { isLoadingMoreArtists = false }
-        let offset = UInt32(artists.count)
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core, libraryPageSize] in
-                try core.listArtists(offset: offset, limit: libraryPageSize)
-            }.value
-            self.artists.append(contentsOf: page.items)
-            self.artistsTotal = page.totalCount
-            serverReachability.noteSuccess()
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .libraryLoad)
-        }
-    }
-
-    /// Refetch the first page of library tracks for the All Tracks tab.
-    /// Called from `refreshLibrary` (inline as an `async let`) on session
-    /// establishment, and available for an explicit retry path later.
-    /// Matches `refreshRecentlyPlayed` in shape — stores items + total.
-    func refreshTracks() async {
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core, libraryInitialPageSize] in
-                try core.listTracks(musicLibraryId: nil, offset: 0, limit: libraryInitialPageSize)
-            }.value
-            self.tracks = page.items
-            self.tracksTotal = page.totalCount
-            serverReachability.noteSuccess()
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .libraryLoad)
-        }
-    }
-
-    /// Fetch the next page of tracks and append to `tracks`. Mirror of
-    /// `loadMoreAlbums` — see its docs for the trigger contract.
-    func loadMoreTracks() async {
-        guard !isLoadingMoreTracks else { return }
-        guard tracksTotal == 0 || tracks.count < Int(tracksTotal) else { return }
-        isLoadingMoreTracks = true
-        defer { isLoadingMoreTracks = false }
-        let offset = UInt32(tracks.count)
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core, libraryPageSize] in
-                try core.listTracks(musicLibraryId: nil, offset: offset, limit: libraryPageSize)
-            }.value
-            self.tracks.append(contentsOf: page.items)
-            self.tracksTotal = page.totalCount
-            serverReachability.noteSuccess()
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .libraryLoad)
-        }
-    }
-
-    /// Resolve (and cache) the `ParentId` to scope playlist queries by.
-    ///
-    /// Calls `core.playlistLibraryId()` once, which hits
-    /// `/Users/{id}/Views` and picks the CollectionFolder whose
-    /// `CollectionType == "playlists"`. On failure we fall back to the
-    /// empty string — Jellyfin's `/Items` endpoint treats an empty
-    /// `ParentId` as "no filter", and the client-side `Path`-based filter
-    /// in `user_playlists` / `public_playlists` still yields a correct set,
-    /// just with more server-side work.
-    func ensurePlaylistLibraryId() async -> String {
-        if let cached = playlistLibraryId { return cached }
-        let resolved: String
-        do {
-            resolved = try await Task.detached(priority: .userInitiated) { [core] in
-                try core.playlistLibraryId()
-            }.value
-        } catch {
-            Log.net.error("ensurePlaylistLibraryId: core.playlistLibraryId() failed (\(error.localizedDescription, privacy: .public)); falling back to empty ParentId.")
-            resolved = ""
-        }
-        playlistLibraryId = resolved
-        return resolved
-    }
-
-    /// Fetch the first page of user-owned playlists for the Library screen's
-    /// Playlists chip. Wired into `refreshLibrary` so the chip is populated
-    /// before the user clicks it. Parallels `loadMoreAlbums` for the error
-    /// / auth / reachability story.
-    ///
-    /// Uses `user_playlists` (user-owned) rather than `public_playlists`. The
-    /// Playlists tab spec (#212) describes "your playlists"; a separate
-    /// "Community" affordance for public playlists is a future concern.
-    func refreshPlaylists() async {
-        let libraryId = await ensurePlaylistLibraryId()
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core, libraryInitialPageSize] in
-                try core.userPlaylists(
-                    playlistLibraryId: libraryId,
-                    offset: 0,
-                    limit: libraryInitialPageSize
-                )
-            }.value
-            self.playlists = page.items
-            self.playlistsTotal = page.totalCount
-            serverReachability.noteSuccess()
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            // Silent-ish: don't clobber the albums/artists error banner if
-            // both fail in the same refresh. The Playlists tab empty state
-            // already explains "nothing to see here" when `playlists` is
-            // empty.
-            Log.net.error("refreshPlaylists failed: \(error.localizedDescription, privacy: .public)")
-        }
-    }
-
-    /// Fetch the next page of playlists and append to `playlists`. Mirror
-    /// of `loadMoreAlbums` — see its docs for the trigger contract.
-    ///
-    /// Server-side total caveat: `user_playlists` filters results client-
-    /// side by `Path`, so `playlistsTotal` is an upper bound on the raw
-    /// server count, not on `playlists.count`. The `<` guard below uses the
-    /// raw total deliberately — stopping at `playlists.count >= total` is
-    /// safe even when the two drift, because the server itself won't return
-    /// more items past its total and we'd bail on an empty page anyway.
-    func loadMorePlaylists() async {
-        guard !isLoadingMorePlaylists else { return }
-        guard playlistsTotal == 0 || playlists.count < Int(playlistsTotal) else { return }
-        isLoadingMorePlaylists = true
-        defer { isLoadingMorePlaylists = false }
-        let libraryId = await ensurePlaylistLibraryId()
-        let offset = UInt32(playlists.count)
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core, libraryPageSize] in
-                try core.userPlaylists(
-                    playlistLibraryId: libraryId,
-                    offset: offset,
-                    limit: libraryPageSize
-                )
-            }.value
-            self.playlists.append(contentsOf: page.items)
-            self.playlistsTotal = page.totalCount
-            serverReachability.noteSuccess()
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            self.errorMessage = LyrebirdErrorPresenter.message(for: error, context: .playlistsLoad)
-        }
-    }
-
-    /// Fetch the user's recently played tracks for the Home screen carousel
-    /// (#206). Passes `nil` for the music library id so the core returns
-    /// tracks across all music libraries the user can see. Failures are
-    /// swallowed silently — an empty carousel is preferable to an error
-    /// banner for a best-effort Home widget.
-    ///
-    /// Stores `totalCount` alongside the page so a future "See all" view can
-    /// expand the carousel without issuing another count query.
-    func refreshRecentlyPlayed() async {
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core, recentlyPlayedInitialPageSize] in
-                try core.recentlyPlayed(
-                    musicLibraryId: nil,
-                    offset: 0,
-                    limit: recentlyPlayedInitialPageSize
-                )
-            }.value
-            self.recentlyPlayed = page.items
-            self.recentlyPlayedTotal = page.totalCount
-        } catch {
-            // Silent fallback — don't surface errors for a secondary widget.
-            _ = handleAuthError(error)
-        }
-    }
-
-    /// Load ALL tracks on a playlist by paging through `playlist_tracks` in
-    /// chunks of `playlistPageSize` until `totalCount` is reached or the
-    /// `playlistSafetyCap` is hit. Returns as soon as any page fails. No UI
-    /// wiring calls this yet (playlist detail screen is #313 et al), but the
-    /// FFI is now paginated so the caller that lands it can rely on "pass
-    /// this a playlist id and get every track". See #125 / #429.
-    func loadAllPlaylistTracks(playlistID: String) async -> [Track] {
-        var all: [Track] = []
-        var offset: UInt32 = 0
-        let limit = playlistPageSize
-        let cap = playlistSafetyCap
-        do {
-            while all.count < cap {
-                let page = try await Task.detached(priority: .userInitiated) { [core] in
-                    try core.playlistTracks(
-                        playlistId: playlistID,
-                        offset: offset,
-                        limit: limit
-                    )
-                }.value
-                all.append(contentsOf: page.items)
-                if page.items.isEmpty { break }
-                if all.count >= Int(page.totalCount) { break }
-                offset = UInt32(all.count)
-            }
-            serverReachability.noteSuccess()
-        } catch {
-            if handleAuthError(error) { return all }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            errorMessage = LyrebirdErrorPresenter.message(for: error, context: .playlistLoad)
-        }
-        return all
-    }
-
-    /// Refresh the Discover "For You" carousel (#249). Until the core exposes
-    /// a real recommendations endpoint (e.g. Jellyfin Items/Suggestions or a
-    /// client-side "artists similar to top-3 played, minus already-played"
-    /// algorithm per research/06-screen-specs.md), this is a best-effort
-    /// stub that mirrors the first 20 recently played tracks so the shelf is
-    /// never empty for an active listener. If `recentlyPlayed` is empty the
-    /// carousel hides itself rather than showing nothing-of-interest.
-    ///
-    /// TODO: replace this stub with a real `core.recommendations(limit: 20)`
-    /// FFI call once it lands. At that point the view layer stays unchanged —
-    /// only the body of this method needs swapping.
-    func refreshForYou() async {
-        // Best-effort fallback: reuse the recently played tracks we already
-        // fetched. Capped at 20 so the carousel stays tight even if the core
-        // later starts returning a longer list.
-        self.forYou = Array(recentlyPlayed.prefix(20))
-    }
-
-    /// Refresh the Discover "Genres to Explore" grid (#250).
-    ///
-    /// Pulls one page of `/MusicGenres` (already filtered server-side to
-    /// genres that carry Audio/Album/Artist items), keeps only those present
-    /// in the library (`song_count > 0`), then ranks them so the *least*
-    /// explored bubble to the top. Jellyfin's `/MusicGenres` projection
-    /// carries no per-genre play count, so we approximate "least-played" with
-    /// ascending `song_count` (the smallest real genres are the ones a user
-    /// is least likely to have worked through), tie-broken by name for a
-    /// stable order. Capped at 8 to fill the 4×2 grid.
-    ///
-    /// Runs the sync `core.genres` FFI off the MainActor (gap pattern #2) and
-    /// marshals the ranked result back. Failures leave the prior grid intact
-    /// rather than blanking the section mid-session.
-    func refreshGenresToExplore() async {
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core] in
-                // limit: 500 matches `resolvedGenreId`'s fetch — /MusicGenres sorts
-                // SortName ascending, so a smaller cap would silently drop
-                // alphabetically-late genres from the ranking pool (the test
-                // library has 254 genres).
-                try core.genres(offset: 0, limit: 500)
-            }.value
-            // Project the FFI genres to plain tuples before ranking. Passing the
-            // `core.Genre` struct directly would force a `LyrebirdCore.Genre`
-            // annotation, which the generated `open class LyrebirdCore` shadows
-            // (module-vs-type name collision — same reason `resolvedGenreId`
-            // tuple-extracts). Tuples keep the ranking helper pure + testable.
-            self.genresToExplore = AppModel.rankGenresToExplore(
-                page.items.map { (id: $0.id, name: $0.name, songCount: $0.songCount) }
-            )
-        } catch {
-            // Auth expiry must surface so the user is routed to re-login, matching
-            // every other refresh path. Non-auth failures leave the prior grid
-            // intact rather than blanking the section mid-session.
-            _ = handleAuthError(error)
-        }
-    }
-
-    /// Pure ranking for the "Genres to Explore" grid (#250), split out from
-    /// the FFI hop so it's unit-testable without a live core. Keeps only
-    /// genres present in the library (`song_count > 0`), ranks ascending by
-    /// `song_count` (least-explored first) with a case-insensitive name
-    /// tiebreaker for stable order, caps at 8 for the 4×2 grid, and carries
-    /// the resolved Jellyfin UUID through into the local `Genre.id`.
-    static func rankGenresToExplore(
-        _ genres: [(id: String, name: String, songCount: UInt32)]
-    ) -> [Genre] {
-        Array(
-            genres
-                .filter { $0.songCount > 0 }
-                .sorted {
-                    if $0.songCount != $1.songCount { return $0.songCount < $1.songCount }
-                    return $0.name.localizedCaseInsensitiveCompare($1.name) == .orderedAscending
-                }
-                .prefix(8)
-                .map { Genre(id: $0.id, name: $0.name) }
-        )
-    }
-
-    /// Refresh the Search "Browse by Genre" tile grid (#247).
-    ///
-    /// Pulls one page of `/MusicGenres` (already filtered server-side to
-    /// genres carrying Audio/Album/Artist items) and ranks them so the
-    /// *biggest* genres lead — the dual of `refreshGenresToExplore`. Jellyfin's
-    /// `/MusicGenres` projection carries `song_count` per genre, so we rank by
-    /// descending count, tie-broken by name for a stable order, and cap at 12
-    /// for the tile grid.
-    ///
-    /// Runs the sync `core.genres` FFI off the MainActor (gap pattern #2) and
-    /// marshals the ranked result back. Failures leave the prior grid intact
-    /// rather than blanking the section mid-session; auth expiry surfaces so
-    /// the user is routed to re-login like every other refresh path.
-    func refreshBrowseGenres() async {
-        do {
-            let page = try await Task.detached(priority: .userInitiated) { [core] in
-                // limit: 500 matches `refreshGenresToExplore` / `resolvedGenreId`
-                // — /MusicGenres sorts SortName ascending, so a smaller cap would
-                // silently drop alphabetically-late genres from the ranking pool
-                // (the test library has 254 genres).
-                try core.genres(offset: 0, limit: 500)
-            }.value
-            // Project the FFI genres to plain tuples before ranking — passing the
-            // `core.Genre` struct directly would force a `LyrebirdCore.Genre`
-            // annotation, which the generated `open class LyrebirdCore` shadows
-            // (module-vs-type name collision). Tuples keep the ranking helper
-            // pure + testable.
-            self.browseGenres = AppModel.rankBrowseGenres(
-                page.items.map { (id: $0.id, name: $0.name, songCount: $0.songCount) }
-            )
-        } catch {
-            _ = handleAuthError(error)
-        }
-    }
-
-    /// Pure ranking for the "Browse by Genre" grid (#247), split out from the
-    /// FFI hop so it's unit-testable without a live core. Keeps only genres
-    /// present in the library (`song_count > 0`), ranks descending by
-    /// `song_count` (biggest first) with a case-insensitive name tiebreaker for
-    /// stable order, caps at 12 for the tile grid, and carries the resolved
-    /// Jellyfin UUID through into the local `Genre.id`.
-    static func rankBrowseGenres(
-        _ genres: [(id: String, name: String, songCount: UInt32)]
-    ) -> [Genre] {
-        Array(
-            genres
-                .filter { $0.songCount > 0 }
-                .sorted {
-                    if $0.songCount != $1.songCount { return $0.songCount > $1.songCount }
-                    return $0.name.localizedCaseInsensitiveCompare($1.name) == .orderedAscending
-                }
-                .prefix(12)
-                .map { Genre(id: $0.id, name: $0.name) }
-        )
-    }
+    var hasAttemptedRestore = false
 
     // MARK: - Items query helpers
-
-    /// Shared helper: build a `GET /Items` request against the user's
-    /// library with the given `sortBy` / `filters`, parse the response,
-    /// and return a typed array of `Album`. Returns an empty array on
-    /// any failure (auth, network, parse) so callers can stay
-    /// conditionally-rendering shelves without an error-banner code path.
-    ///
-    /// TODO(core-#465): replace with a typed `core.items_query()` builder
-    ///   once that FFI exists. This function's surface lines up
-    ///   deliberately with the shape that builder will expose.
-    func fetchAlbumsViaItemsQuery(
-        sortBy: String,
-        filters: String?,
-        limit: UInt32,
-        extraFields: [String],
-        minDateLastSaved: String?
-    ) async -> [Album] {
-        guard let request = buildItemsQuery(
-            includeItemTypes: "MusicAlbum",
-            sortBy: sortBy,
-            sortOrder: "Descending",
-            filters: filters,
-            limit: limit,
-            extraFields: extraFields,
-            minDateLastSaved: minDateLastSaved,
-            parentId: nil
-        ) else { return [] }
-        do {
-            let (data, resp) = try await URLSession.shared.data(for: request)
-            if let http = resp as? HTTPURLResponse, http.statusCode == 401 {
-                markAuthExpired()
-                return []
-            }
-            return Self.parseAlbumsFromItems(data: data)
-        } catch {
-            Log.net.error("fetchAlbumsViaItemsQuery failed: \(error.localizedDescription, privacy: .public)")
-            return []
-        }
-    }
-
-    /// Like `fetchAlbumsViaItemsQuery` but also returns a map from album
-    /// id to the server-reported `UserData.PlayCount` so "N plays" can
-    /// render on the Quick Picks tile.
-    func fetchAlbumsWithPlayCounts(
-        sortBy: String,
-        filters: String?,
-        limit: UInt32,
-        minDateLastSaved: String?
-    ) async -> ([Album], [String: UInt32]) {
-        guard let request = buildItemsQuery(
-            includeItemTypes: "MusicAlbum",
-            sortBy: sortBy,
-            sortOrder: "Descending",
-            filters: filters,
-            limit: limit,
-            extraFields: [],
-            minDateLastSaved: minDateLastSaved,
-            parentId: nil
-        ) else { return ([], [:]) }
-        do {
-            let (data, resp) = try await URLSession.shared.data(for: request)
-            if let http = resp as? HTTPURLResponse, http.statusCode == 401 {
-                markAuthExpired()
-                return ([], [:])
-            }
-            return Self.parseAlbumsWithPlayCounts(data: data)
-        } catch {
-            Log.net.error("fetchAlbumsWithPlayCounts failed: \(error.localizedDescription, privacy: .public)")
-            return ([], [:])
-        }
-    }
-
-    /// Fetch Recently Added via `/Users/{id}/Items/Latest`. Returns both
-    /// the album array and a per-album `DateCreated` map (used by the NEW
-    /// badge on `RecentlyAddedTile`).
-    func fetchLatestAlbumsWithDates(limit: UInt32) async -> ([Album], [String: Date]) {
-        guard let session = session,
-              let baseURL = URL(string: session.server.url),
-              let authHeader = try? core.authHeader()
-        else { return ([], [:]) }
-        let userId = session.user.id
-        var comps = URLComponents(
-            url: baseURL.appendingPathComponent("Users/\(userId)/Items/Latest"),
-            resolvingAgainstBaseURL: false
-        )
-        comps?.queryItems = [
-            URLQueryItem(name: "IncludeItemTypes", value: "MusicAlbum"),
-            URLQueryItem(name: "Limit", value: String(limit)),
-            URLQueryItem(name: "GroupItems", value: "true"),
-            URLQueryItem(
-                name: "Fields",
-                value: "Genres,ProductionYear,DateCreated,ChildCount,PrimaryImageAspectRatio"
-            ),
-        ]
-        guard let url = comps?.url else { return ([], [:]) }
-        var request = URLRequest(url: url)
-        request.setValue(authHeader, forHTTPHeaderField: "Authorization")
-        request.setValue("application/json", forHTTPHeaderField: "Accept")
-        do {
-            let (data, resp) = try await URLSession.shared.data(for: request)
-            if let http = resp as? HTTPURLResponse, http.statusCode == 401 {
-                markAuthExpired()
-                return ([], [:])
-            }
-            // `/Items/Latest` returns a bare array, not the
-            // `{Items, TotalRecordCount}` wrapper — parse accordingly.
-            return Self.parseLatestAlbumsWithDates(data: data)
-        } catch {
-            Log.net.error("fetchLatestAlbumsWithDates failed: \(error.localizedDescription, privacy: .public)")
-            return ([], [:])
-        }
-    }
-
-    /// Fetch up to `limit` favorited audio tracks. Backs the
-    /// "Shuffle All Favorites" CTA on the Home Favorites header (#55).
-    func fetchFavoriteTracks(limit: UInt32) async -> [Track] {
-        guard let request = buildItemsQuery(
-            includeItemTypes: "Audio",
-            sortBy: "Random",
-            sortOrder: "Ascending",
-            filters: "IsFavorite",
-            limit: limit,
-            extraFields: [],
-            minDateLastSaved: nil,
-            parentId: nil
-        ) else { return [] }
-        do {
-            let (data, resp) = try await URLSession.shared.data(for: request)
-            if let http = resp as? HTTPURLResponse, http.statusCode == 401 {
-                markAuthExpired()
-                return []
-            }
-            return Self.parseTracksFromItems(data: data)
-        } catch {
-            Log.net.error("fetchFavoriteTracks failed: \(error.localizedDescription, privacy: .public)")
-            return []
-        }
-    }
 
     /// Per-session cache for `artistDetail`. Cleared on `logout()` / `forgetToken()`.
     var artistDetailCache: [String: ArtistDetail] = [:]
@@ -1843,12 +872,12 @@ final class AppModel {
     /// Stored as the opaque `AmbientPalette.encoded` string rather than the
     /// `Color`-bearing struct so the entry is a plain value type and survives
     /// any future move to a persistent cache without a bespoke codable surface.
-    private var ambientPaletteCache: [String: String] = [:]
+    var ambientPaletteCache: [String: String] = [:]
 
     /// In-flight palette samples, so two `NowPlayingView.task` invocations for
     /// the same id (e.g. a fast tab toggle) coalesce onto one Nuke decode +
     /// CIAreaAverage pass instead of racing two.
-    private var ambientPaletteTasks: [String: Task<AmbientPalette?, Never>] = [:]
+    var ambientPaletteTasks: [String: Task<AmbientPalette?, Never>] = [:]
 
     /// Resolve the ambient palette for a now-playing item — cache-first, then
     /// a single off-main Nuke decode + Core Image average (`PaletteSampler`).
@@ -1938,312 +967,6 @@ final class AppModel {
     /// drive the "played" glyph in track rows and to compute the toggle
     /// target state in `toggleMarkPlayed(tracks:)`. See #133.
     var playedById: [String: Bool] = [:]
-
-    /// Toggle the favorite flag for an album on the Jellyfin server. Reads
-    /// the current state from `favoriteById` (falling back to `false` on a
-    /// cold start) and calls the opposite side of `set_favorite` /
-    /// `unset_favorite` on the core. The returned [`FavoriteState`] is the
-    /// server's authoritative answer and is written back to `favoriteById`
-    /// so the heart glyph reflects the saved state.
-    ///
-    /// Errors surface the generic `errorMessage` banner — a failed toggle is
-    /// rare enough that swallowing it would hide real trouble (token
-    /// revoked, network flapping), but not so load-bearing that we want a
-    /// modal.
-    func toggleFavorite(album: Album) {
-        Task { await setFavorite(itemId: album.id, enabled: !isFavorite(album: album)) }
-    }
-
-    /// Toggle the favorite flag for a track. Same contract as
-    /// `toggleFavorite(album:)` — see its doc for the state-cache semantics.
-    func toggleFavorite(track: Track) {
-        Task { await setFavorite(itemId: track.id, enabled: !isFavorite(track: track)) }
-    }
-
-    /// Check the local favorite-state cache. Returns `false` when the item
-    /// hasn't been toggled this session AND no snapshot is available — the
-    /// snapshot-aware overloads `isFavorite(track:)` / `isFavorite(album:)` /
-    /// `isFavorite(artist:)` are preferred at call sites that have a model
-    /// object on hand because they read the server-authoritative
-    /// `userData.isFavorite` from the snapshot when the cache is cold.
-    func isFavorite(id: String) -> Bool {
-        favoriteById[id] ?? false
-    }
-
-    /// Snapshot-aware favorite check for tracks. Reads the in-memory cache
-    /// first (toggled-this-session is authoritative), then falls back to
-    /// the server-authoritative `track.userData?.isFavorite` projection,
-    /// then to the legacy top-level `track.isFavorite` mirror. This is the
-    /// preferred read for any heart-glyph UI that has the `Track` value on
-    /// hand: it shows the correct state on first paint for already-favorited
-    /// tracks (the cache-only `isFavorite(id:)` returns `false` until the
-    /// user toggles, which makes the next tap appear to no-op against the
-    /// server). See the rc6 favorite-cache seeding fix.
-    func isFavorite(track: Track) -> Bool {
-        if let cached = favoriteById[track.id] { return cached }
-        if let userFav = track.userData?.isFavorite { return userFav }
-        return track.isFavorite
-    }
-
-    /// Snapshot-aware favorite check for albums. Mirrors
-    /// `isFavorite(track:)` — falls back to `album.userData?.isFavorite`
-    /// when the cache is cold so the album-detail heart shows the correct
-    /// state on first paint. `Album` has no legacy top-level `isFavorite`
-    /// mirror so the final fallback is `false`.
-    func isFavorite(album: Album) -> Bool {
-        if let cached = favoriteById[album.id] { return cached }
-        return album.userData?.isFavorite ?? false
-    }
-
-    /// Snapshot-aware favorite check for playlists. Mirrors
-    /// `isFavorite(album:)` — falls back to `playlist.userData?.isFavorite`
-    /// when the cache is cold so the playlist-detail heart shows the correct
-    /// state on first paint.
-    func isFavorite(playlist: Playlist) -> Bool {
-        if let cached = favoriteById[playlist.id] { return cached }
-        return playlist.userData?.isFavorite ?? false
-    }
-
-    /// Snapshot-aware favorite check for artists. Mirrors
-    /// `isFavorite(track:)` — falls back to `artist.userData?.isFavorite`.
-    func isFavorite(artist: Artist) -> Bool {
-        if let cached = favoriteById[artist.id] { return cached }
-        return artist.userData?.isFavorite ?? false
-    }
-
-    /// Internal helper — hits `set_favorite` / `unset_favorite` on the core
-    /// and mirrors the server's answer into `favoriteById`. `internal` (not
-    /// `private`) so the `toggleFavorite(...)` wrappers — some now in
-    /// `AppModel+*` extension files (e.g. `AppModel+Playlists.swift`) — can
-    /// route through it; the `toggleFavorite(...)` API stays the preferred
-    /// entry point so the desired-state boolean is always computed at the
-    /// call site.
-    func setFavorite(itemId: String, enabled: Bool) async {
-        Log.tracks.info("setFavorite item=\(itemId, privacy: .public) target=\(enabled, privacy: .public)")
-        do {
-            let state = try await Task.detached(priority: .userInitiated) { [core] in
-                if enabled {
-                    return try core.setFavorite(itemId: itemId)
-                } else {
-                    return try core.unsetFavorite(itemId: itemId)
-                }
-            }.value
-            Log.tracks.info("setFavorite ok item=\(itemId, privacy: .public) server=\(state.isFavorite, privacy: .public)")
-            favoriteById[itemId] = state.isFavorite
-            favoriteChangeToken &+= 1
-            serverReachability.noteSuccess()
-            // If the app UI just favorited the currently-playing track, sync
-            // Control Center's like indicator — only the remote-command path
-            // self-refreshes otherwise (#460).
-            if status.currentTrack?.id == itemId {
-                mediaSession.refreshTransportState()
-            }
-        } catch {
-            Log.tracks.error("setFavorite failed item=\(itemId, privacy: .public) error=\(error.localizedDescription, privacy: .public)")
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            errorMessage = LyrebirdErrorPresenter.message(for: error, context: .favorite)
-        }
-    }
-
-    /// Read the locally-cached played state for an item id. Mirrors
-    /// `isFavorite(id:)`; used by row views + `toggleMarkPlayed(tracks:)`
-    /// to compute the target state for a multi-select toggle. See #133.
-    func isPlayed(id: String) -> Bool {
-        playedById[id] ?? false
-    }
-
-    /// Internal helper — hits `mark_played` / `mark_unplayed` on the core
-    /// and mirrors the server's answer (full `UserItemData`) into
-    /// `playedById`. Mirrors `setFavorite(itemId:enabled:)` in shape so a
-    /// single-item toggle has a single failure path. See #133.
-    func setPlayed(itemId: String, played: Bool) async {
-        do {
-            let state = try await Task.detached(priority: .userInitiated) { [core] in
-                try core.setPlayed(itemId: itemId, played: played)
-            }.value
-            playedById[itemId] = state.played
-            serverReachability.noteSuccess()
-        } catch {
-            if handleAuthError(error) { return }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            errorMessage = LyrebirdErrorPresenter.message(for: error, context: .markPlayed)
-        }
-    }
-
-    /// Download every track on the album for offline playback (#819).
-    /// Resolves the album's tracks (cache-or-fetch), then hands them to the
-    /// shared `downloadTracks` pipeline. Gated behind `supportsDownloads`; a
-    /// no-op when the feature is dormant.
-    func enqueueDownload(album: Album) {
-        guard supportsDownloads else { return }
-        Task {
-            let tracks = await loadTracks(forAlbum: album.id)
-            guard !tracks.isEmpty else { return }
-            await downloadTracks(tracks)
-        }
-    }
-
-    /// Append a batch of tracks to an existing playlist via `add_to_playlist`
-    /// on the core. Used by the album detail popover (#222) and any other
-    /// caller that has already resolved a target playlist. Returns `true` on
-    /// success so UI can dismiss the popover / show a confirmation tick.
-    ///
-    /// Errors surface on `errorMessage` rather than throwing so the popover
-    /// can stay presentation-only. An empty `trackIds` short-circuits before
-    /// the FFI hop since the server would reject it anyway.
-    @discardableResult
-    func addToPlaylist(trackIds: [String], playlistId: String) async -> Bool {
-        guard !trackIds.isEmpty else { return false }
-        do {
-            try await Task.detached(priority: .userInitiated) { [core] in
-                try core.addToPlaylist(playlistId: playlistId, itemIds: trackIds, position: nil)
-            }.value
-            serverReachability.noteSuccess()
-            return true
-        } catch {
-            if handleAuthError(error) { return false }
-            if ServerReachability.shouldCount(error: error) {
-                serverReachability.noteFailure()
-            }
-            errorMessage = LyrebirdErrorPresenter.message(for: error, context: .playlistAdd)
-            return false
-        }
-    }
-
-    /// Fetch the album's hydrated detail fields (label, premiere date, and
-    /// aggregated People credits) via `fetch_item`. Returned as a compact
-    /// [`AlbumDetail`] value type so the view layer can render the
-    /// liner-note credits section (#65) without a second parse pass.
-    ///
-    /// Silent on errors: the liner-note section degrades to whatever fields
-    /// are present on the cached `Album` so a 404 or a stripped-down server
-    /// doesn't take down the whole detail page.
-    func loadAlbumDetail(albumId: String) async -> AlbumDetail {
-        do {
-            let json = try await Task.detached(priority: .userInitiated) { [core] in
-                try core.fetchItem(
-                    itemId: albumId,
-                    fields: ["People", "Studios", "PremiereDate", "DateCreated", "ProductionYear", "Overview"]
-                )
-            }.value
-            return Self.parseAlbumDetail(from: json)
-        } catch {
-            _ = handleAuthError(error)
-            return AlbumDetail(label: nil, releaseDate: nil, people: [], overview: nil)
-        }
-    }
-
-    /// Parse the subset of the album item JSON that the liner-note section
-    /// cares about. Static + internal so tests can hit it without wiring
-    /// the full model. Missing fields become `nil`; the parser never throws.
-    static func parseAlbumDetail(from json: String) -> AlbumDetail {
-        guard let data = json.data(using: .utf8),
-              let root = try? JSONSerialization.jsonObject(with: data) as? [String: Any]
-        else {
-            return AlbumDetail(label: nil, releaseDate: nil, people: [], overview: nil)
-        }
-
-        // Jellyfin ships `Studios` as an array of `{ Name, Id }` objects. Pick
-        // the first non-empty label — servers with multiple labels tend to
-        // list the primary one first.
-        let label: String? = {
-            guard let studios = root["Studios"] as? [[String: Any]] else { return nil }
-            for entry in studios {
-                if let name = entry["Name"] as? String {
-                    let trimmed = name.trimmingCharacters(in: .whitespaces)
-                    if !trimmed.isEmpty { return trimmed }
-                }
-            }
-            return nil
-        }()
-
-        // `PremiereDate` is an ISO 8601 string; fall back to `DateCreated`
-        // if absent. We only keep the yyyy-MM-dd portion since the hero
-        // already shows the year and the liner-note section wants "Released
-        // 19 Apr 2013".
-        let releaseDate: Date? = {
-            let iso = ISO8601DateFormatter()
-            iso.formatOptions = [.withInternetDateTime, .withFractionalSeconds]
-            for key in ["PremiereDate", "DateCreated"] {
-                if let raw = root[key] as? String, !raw.isEmpty {
-                    if let d = iso.date(from: raw) { return d }
-                    iso.formatOptions = [.withInternetDateTime]
-                    if let d = iso.date(from: raw) { return d }
-                }
-            }
-            return nil
-        }()
-
-        let people: [Person] = {
-            guard let raw = root["People"] as? [[String: Any]] else { return [] }
-            return raw.compactMap { entry in
-                let name = (entry["Name"] as? String)?.trimmingCharacters(in: .whitespaces) ?? ""
-                let type = (entry["Type"] as? String) ?? ""
-                let rawId = (entry["Id"] as? String)?.trimmingCharacters(in: .whitespaces) ?? ""
-                let id = rawId.isEmpty ? nil : rawId
-                guard !name.isEmpty else { return nil }
-                return Person(name: name, type: type, id: id)
-            }
-        }()
-
-        // Editorial blurb (#68). Kept raw here — the album detail view runs it
-        // through the same HTML strip as the artist bio before display.
-        // Whitespace-only collapses to `nil` so the "About this album" section
-        // never renders an empty shell.
-        let overview: String? = {
-            guard let raw = root["Overview"] as? String else { return nil }
-            let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
-            return trimmed.isEmpty ? nil : raw
-        }()
-
-        return AlbumDetail(label: label, releaseDate: releaseDate, people: people, overview: overview)
-    }
-
-    /// Navigate to the artist detail screen for this album's artist, if known.
-    func goToArtist(album: Album) {
-        guard let artistID = album.artistId else { return }
-        navPath.append(Route.artist(artistID))
-    }
-
-    /// Navigate to the album's own detail screen. Used when the menu is
-    /// invoked from a surface other than the album detail itself (e.g. a
-    /// track row that links back to its album).
-    func goToAlbum(album: Album) {
-        navPath.append(Route.album(album.id))
-    }
-
-    /// Kick off an Instant Mix ("album radio") seeded by this album.
-    func startAlbumRadio(album: Album) {
-        playInstantMix(seedId: album.id)
-    }
-
-    /// Mark the album as played server-side. Matches Jellyfin web's
-    /// "Mark Played" affordance: the album's `UserData.Played` flips and
-    /// `LastPlayedDate` stamps, but per-track `PlayCount` is **not**
-    /// incremented (Jellyfin doesn't cascade the operation to children).
-    /// `playedById[album.id]` flips immediately for the optimistic glyph
-    /// flip, then reconciles with the server's authoritative response. See #133.
-    func markAllAsPlayed(album: Album) {
-        let target = !isPlayed(id: album.id)
-        playedById[album.id] = target
-        Task { await setPlayed(itemId: album.id, played: target) }
-    }
-
-    /// Append every track on the album to a user-picked playlist. Loads
-    /// the tracklist (cached) then routes through the async
-    /// `addToPlaylist(trackIds:playlistId:)` path.
-    func addAlbumToPlaylist(album: Album, playlist: Playlist) {
-        Task {
-            let tracks = await loadTracks(forAlbum: album.id)
-            guard !tracks.isEmpty else { return }
-            await addToPlaylist(trackIds: tracks.map(\.id), playlistId: playlist.id)
-        }
-    }
 
     // MARK: - Track info sheet
